requirements.txt
flask>=3.0.2
openai>=1.12.0
selenium>=4.18.1
pandas>=2.2.0
openpyxl>=3.1.2
numpy>=1.26.0
apscheduler>=3.10.4
gunicorn>=21.2.0
python-dotenv>=1.0.0
PyJWT>=2.8.0 